requirements.txt
appdirs==1.4.3
asgiref==3.2.3
distlib==0.3.0
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.2
django-heroku==0.3.1
djangorestframework==3.11.0
filelock==3.0.12
gunicorn==20.0.4
node==0.9.25
odict==1.7.0
Pillow==7.0.0
plumber==1.5
psycopg2==2.8.5
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
static3==0.7.0
unicorn==1.0.1
virtualenv==20.0.8
whitenoise==5.0.1
zope.component==4.6.1
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==5.0.1
zope.interface==5.1.0
zope.lifecycleevent==4.3
zope.proxy==4.3.5
